Cliente POST

+++ backend/src/main/java/com/vendaspedidos/entities/Cidade.java
@@ -17,7 +17,7 @@ public class Cidade  implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 	@ManyToOne
@@ -27,17 +27,17 @@ public class Cidade  implements Serializable{
 	public Cidade() {
 	}
 
-	public Cidade(Long id, String nome, Estado estado) {
+	public Cidade(Integer id, String nome, Estado estado) {
 		this.id = id;
 		this.nome = nome;
 		this.estado = estado;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
